+++ StripeFinancialConnections/StripeFinancialConnections/Source/API Bindings/FinancialConnectionsAPIClient.swift
@@ -294,20 +294,26 @@ extension FinancialConnectionsAPIClient: FinancialConnectionsAPI {
         clientSecret: String,
         returnURL: String?
     ) -> Future<FinancialConnectionsSynchronize> {
-        let parameters: [String: Any] = [
+        var parameters: [String: Any] = [
             "expand": ["manifest.active_auth_session"],
             "client_secret": clientSecret,
-            "mobile": {
-                var mobileParameters: [String: Any] = [
-                    "fullscreen": true,
-                    "hide_close_button": true,
-                    "forced_authflow_version": "v3",
-                ]
-                mobileParameters["app_return_url"] = returnURL
-                return mobileParameters
-            }(),
             "locale": Locale.current.toLanguageTag(),
         ]
+
+        var mobileParameters: [String: Any] = [
+            "fullscreen": true,
+            "hide_close_button": true,
+            "forced_authflow_version": "v3",
+        ]
+        mobileParameters["app_return_url"] = returnURL
+
+        let attest = backingAPIClient.stripeAttest
+        if attest.isSupported {
+            mobileParameters["supports_app_verification"] = true
+            mobileParameters["verified_app_id"] = Bundle.main.bundleIdentifier
+        }
+        parameters["mobile"] = mobileParameters
+
         return self.post(
             resource: "financial_connections/sessions/synchronize",
             parameters: parameters,
@@ -1108,6 +1114,25 @@ extension FinancialConnectionsAPIClient: FinancialConnectionsAPI {
                 )
             }
     }
+    
+    func updateAvailableIncentives(
+        consumerSessionClientSecret: String,
+        sessionID: String
+    ) -> Future<AvailableIncentives> {
+        let parameters: [String: Any] = [
+            "request_surface": requestSurface,
+            "credentials": [
+                "consumer_session_client_secret": consumerSessionClientSecret
+            ],
+            "session_id": sessionID,
+        ]
+
+        return post(
+            resource: APIEndpointAvailableIncentives,
+            parameters: parameters,
+            useConsumerPublishableKeyIfNeeded: true
+        )
+    }
 }
 
 private let APIEndpointListAccounts = "link_account_sessions/list_accounts"
@@ -1142,3 +1167,4 @@ private let APIEndpointAttachLinkConsumerToLinkAccountSession = "consumers/attac
 private let APIEndpointPaymentDetails = "consumers/payment_details"
 private let APIEndpointSharePaymentDetails = "consumers/payment_details/share"
 private let APIEndpointPaymentMethods = "payment_methods"
+private let APIEndpointAvailableIncentives = "consumers/incentives/update_available"